refactor: rename ProposalActionRoleInput file
