+++ src/proposals/proposal-actions/models/proposal-action-role-input.ts
